refactor: use integer values instead of doubles for block location
Double values are not needed. Furthermore, chances exist that it may lead to confusions and hence bugs.
Following the rule "use strictly only what it is needed".

+++ src/bukkit/bukkit-patch-adapter/src/main/java/fr/djaytan/minecraft/jobsreborn/patchplacebreak/bukkit/adapter/converter/BlockFaceConverter.java
@@ -34,9 +34,9 @@ public class BlockFaceConverter implements UnidirectionalConverter<BlockFace, Ve
 
   @Override
   public @NonNull Vector convert(@NonNull BlockFace blockFace) {
-    double modX = blockFace.getModX();
-    double modY = blockFace.getModY();
-    double modZ = blockFace.getModZ();
+    int modX = blockFace.getModX();
+    int modY = blockFace.getModY();
+    int modZ = blockFace.getModZ();
     return Vector.of(modX, modY, modZ);
   }
 }
