+++ src/patch-place-break/patch-place-break-api/src/main/java/fr/djaytan/minecraft/jobsreborn/patchplacebreak/api/entities/Vector.java
@@ -29,7 +29,7 @@ import lombok.Value;
 @Value(staticConstructor = "of")
 public class Vector {
 
-  double modX;
-  double modY;
-  double modZ;
+  int modX;
+  int modY;
+  int modZ;
 }
